+++ package.json
@@ -1,7 +1,7 @@
 {
   "private": true,
   "name": "webchatgpt",
-  "version": "3.2.7",
+  "version": "3.3.1",
   "license": "MIT",
   "scripts": {
     "build-dev": "node build.mjs",
