+++ src/background/bg.ts
@@ -1,5 +1,5 @@
 import Browser from 'webextension-polyfill'
-import { getHtml } from 'src/content-scripts/ddg_search'
+import { getHtml } from 'src/content-scripts/web_search'
 import { getWebpageTitleAndText } from 'src/content-scripts/api'
 
 
@@ -16,7 +16,6 @@ function openChatGPTWebpage() {
 
 if (manifest_version == 2) {
     Browser.browserAction.onClicked.addListener(openChatGPTWebpage)
-    update_origin_for_ddg_in_firefox()
 } else {
     Browser.action.onClicked.addListener(openChatGPTWebpage)
 }
@@ -48,48 +47,3 @@ Browser.runtime.onMessage.addListener((message) => {
         return getWebpageTitleAndText(message.url, message.html)
     }
 })
-
-// Firefox does not support declarativeNetRequest.updateDynamicRules yet
-Browser.declarativeNetRequest.updateDynamicRules({
-    addRules: [
-        {
-            id: 1,
-            priority: 1,
-            action: {
-                type: "modifyHeaders",
-                requestHeaders: [
-                    {
-                        header: "Origin",
-                        operation: "set",
-                        value: "https://lite.duckduckgo.com"
-                    },
-                ],
-            },
-
-            condition: {
-                urlFilter: "https://lite.duckduckgo.com/*",
-                resourceTypes: ["xmlhttprequest"],
-            },
-        },
-    ],
-    removeRuleIds: [1],
-})
-
-function update_origin_for_ddg_in_firefox() {
-    Browser.webRequest.onBeforeSendHeaders.addListener(
-        (details) => {
-            if (!details.requestHeaders) return
-            for (let i = 0; i < details.requestHeaders.length; ++i) {
-                if (details.requestHeaders[i].name === 'Origin')
-                    details.requestHeaders[i].value = "https://lite.duckduckgo.com"
-            }
-
-            return {
-                requestHeaders: details.requestHeaders
-            }
-        }, {
-        urls: ["https://lite.duckduckgo.com/*"],
-    },
-        ["blocking", "requestHeaders"]
-    )
-}
