+++ src/util/localization.ts
@@ -55,7 +55,6 @@ export const localizationKeys = {
         supportMe: 'support_me',
         chooseLanguage: 'choose_language',
         textareaPlaceholder: 'textarea_placeholder',
-        youCanUseDuckDuckGoBangs: 'you_can_use_duckduckgo_bangs',
     },
     slashCommandsMenu: {
         siteCommandDescription: 'site_command_description',
