+++ src/util/promptManager.ts
@@ -2,7 +2,7 @@ import Browser from "webextension-polyfill"
 import { v4 as uuidv4 } from 'uuid'
 import { getCurrentLanguageName, getLocaleLanguage, getTranslation, localizationKeys } from "./localization"
 import { getUserConfig } from "./userConfig"
-import { SearchResult } from "src/content-scripts/ddg_search"
+import { SearchResult } from "src/content-scripts/web_search"
 
 export const SAVED_PROMPTS_KEY = 'saved_prompts'
 export const SAVED_PROMPTS_MOVED_KEY = 'saved_prompts_moved_to_local'
