+++ src/util/timePeriodOptions.json
@@ -2,6 +2,5 @@
   { "value": "", "label": "Any time" },
   { "value": "d", "label": "Past day" },
   { "value": "w", "label": "Past week" },
-  { "value": "m", "label": "Past month" },
-  { "value": "y", "label": "Past year" }
+  { "value": "m", "label": "Past month" }
 ]
